refactor: move PromotionCard.tsx to subdirectory promotion-card
