+++ src/app/main-page/promotion-card/promotion-card-styles.ts
